requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
sqlalchemy==2.0.0
pyjwt==2.8.0
python-multipart==0.0.6
bcrypt==4.0.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pillow==10.0.0
pydantic>=1.7.4,<3.0.0,!=2.0.0,!=2.0.1
email-validator
